se ajustaron alertas y css

+++ src/components/Login.js
@@ -138,8 +138,8 @@ export const Login = (onNavigate) => {
           <p>${task.description}<p>
           <p class='displayName'> 👨🏽‍🍳${task.displayName}</p>
           <button class='btn-like' data-id='${doc.id}'>🖤</button>
-          <button class='btn-delete' data-id='${doc.id}'>Delete</button>
-          <button class='btn-edit' data-id='${doc.id}'>Edit</button>
+          <button class='btn-delete' data-id='${doc.id}'>Eliminar</button>
+          <button class='btn-edit' data-id='${doc.id}'>Editar</button>
         </div>
           `;
     });
@@ -149,18 +149,39 @@ export const Login = (onNavigate) => {
     // console.log(btnsDelete);
     btnsDelete.forEach((btn) => {
       btn.addEventListener('click', ({ target: { dataset } }) => {
-        deleteTask(dataset.id);
+        Swal.fire({
+          title: '¿Esta seguro que desea eliminar esta publicacion?',
+          icon: 'warning',
+          showCancelButton: true,
+          width: 300,
+          Height: 100,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Si',
+        }).then((result) => {
+          if (result.isConfirmed) {
+            deleteTask(dataset.id);
+            Swal.fire(
+              'Publicacion eliminada.',
+              '',
+              'success',
+            );
+          } else {
+            onNavigate('/login');
+          }
+        });
       });
     });
-    // boton likes
-    const btnLike = taskContainer.querySelectorAll('.btn-like');
-    // console.log(btnLike);
-    btnLike.forEach((btn) => {
-      btn.addEventListener('click', ({ target: { dataset } }) => {
-        addLikePost(dataset.id, usuarioId);
-        console.log(btnLike);
-      });
-    });
+
+    // botton likes
+    // const btnLike = taskContainer.querySelectorAll('.btn-like');
+    /// / console.log(btnLike);
+    // btnLike.forEach((btn) => {
+    //  btn.addEventListener('click', ({ target: { dataset } }) => {
+    //    addLikePost(dataset.id, usuarioId);
+    //    console.log(btnLike);
+    //  });
+    // });
 
     const btnsEdit = taskContainer.querySelectorAll('.btn-edit');
